% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_Diag.R
\name{cus_reg_diag}
\alias{cus_reg_diag}
\title{All-in-One function for regression diagnostics.}
\usage{
cus_reg_diag(
  lm,
  summaryOnly = TRUE,
  interactionYN = FALSE,
  sig_level = 0.05,
  cutoff_vif = 5
)
}
\arguments{
\item{lm}{Please enter lm object with the following notation: lm(data = data, Y~Xi+Xn)}

\item{interactionYN}{TRUE or FALSE, are interactions included? Will calculate GVIF instead of VIF.}

\item{sig_level}{A value between 0 and 1. If a test's p-value is lager than sig_level, the assumption is met.}

\item{cutiff_vif}{Cutoof-VIF threshold for identifying multicoliniearity.}
}
\description{
Computes statistical tests and plots for linear regression diagnostics.
Plots are also generated for a multivariate normal comparison set.
}
