% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_Diag.R
\name{lm_diagnostic}
\alias{lm_diagnostic}
\title{All-in-One function for regression diagnostics.}
\usage{
lm_diagnostic(
  lm,
  summaryOnly = FALSE,
  interactionYN = FALSE,
  sig_level = 0.05,
  cutoff_vif = 5,
  seed = 123,
  rMarkdown = FALSE
)
}
\arguments{
\item{lm}{Please enter lm object with the following notation: lm(data = data, Y~Xi+Xn)}

\item{summaryOnly}{Only shows the statistical test results, omitting plots.}

\item{interactionYN}{TRUE or FALSE, are interactions included? Will calculate GVIF instead of VIF.}

\item{sig_level}{A value between 0 and 1. If a test's p-value is lager than sig_level, the assumption is met.}

\item{cutoff_vif}{Cutoof-VIF threshold for identifying multicoliniearity.}

\item{seed}{Set seed for reproducible number generation.}
}
\description{
Computes statistical tests and plots for linear regression diagnostics.
Plots are also generated for a multivariate normal comparison set.
}
